notification-manager-MVCtest-loginPage/notification-manager-MVCtest/requirements.txt
flask
firebase_admin
flask-SQLAlchemy
flask_migrate
flask_login
pymysql